+++ Tests/TuistEnvKitTests/Versions/Mocks/MockVersionsController.swift
@@ -1,6 +1,7 @@
 import Foundation
 import TSCBasic
 import struct TSCUtility.Version
+import TuistSupport
 import TuistSupportTesting
 @testable import TuistEnvKit
